requirements.txt
pandas
numpy
plotly
streamlit
pandasai
pyYAML
langchain_google_genai
google-generativeai
python-dotenv